# backend/internal/interfaces/http/v1/handlers/memory.go
// Package handlers implements HTTP request handlers following Clean Architecture principles.
//
// PURPOSE: Serves as the Interface Adapter layer that transforms HTTP requests into
// application use cases and converts application responses back to HTTP format.
// This is the entry point for all REST API operations in the Brain2 system.
//
// CLEAN ARCHITECTURE ROLE: This layer handles external communication concerns:
//   • Request/Response transformation and validation
//   • HTTP status code mapping from domain errors
//   • Authentication and authorization enforcement
//   • API versioning and backward compatibility
//   • Request tracing and observability integration
//
// KEY HANDLERS:
//   • MemoryHandler: CRUD operations for memory nodes and graph queries
//   • CategoryHandler: Category management and auto-categorization
//   • HealthHandler: System health checks and readiness probes
//
// DESIGN PRINCIPLES:
//   • Thin Layer: No business logic, only coordination and transformation
//   • Dependency Injection: All dependencies injected via constructor
//   • Error Handling: Consistent error responses across all endpoints
//   • Observability: Request tracing, metrics, and structured logging
//
// This package ensures HTTP concerns remain separate from business logic,
// enabling easy testing and potential protocol changes in the future.
package handlers

import (
	"crypto/sha256"
	"encoding/hex"
	"encoding/json"
	"fmt"
	"log"
	"net/http"
	"strconv"
	"time"

	"brain2-backend/internal/application/commands"
	"brain2-backend/internal/application/queries"
	"brain2-backend/internal/application/services"
	"brain2-backend/pkg/api"

	"github.com/aws/aws-sdk-go-v2/aws"
	"github.com/aws/aws-sdk-go-v2/service/eventbridge"
	"github.com/aws/aws-sdk-go-v2/service/eventbridge/types"
	"github.com/go-chi/chi/v5"
)

// MemoryHandler handles memory-related HTTP requests with CQRS services.
type MemoryHandler struct {
	// CQRS services for clean separation of concerns
	nodeService       *services.NodeService      // Write operations (commands)
	nodeQueryService  *queries.NodeQueryService  // Read operations (queries)
	graphQueryService *queries.GraphQueryService // Graph operations (queries)

	// Infrastructure dependencies
	eventBridgeClient *eventbridge.Client
	container         interface {
		IsPostColdStartRequest() bool
		GetTimeSinceColdStart() time.Duration
	}
}

// ColdStartContainer interface for cold start detection
type ColdStartContainer interface {
	IsPostColdStartRequest() bool
	GetTimeSinceColdStart() time.Duration
}

// NewMemoryHandler creates a new memory handler with CQRS services.
func NewMemoryHandler(
	nodeService *services.NodeService,
	nodeQueryService *queries.NodeQueryService,
	graphQueryService *queries.GraphQueryService,
	eventBridgeClient *eventbridge.Client,
	container ColdStartContainer,
) *MemoryHandler {
	return &MemoryHandler{
		nodeService:       nodeService,
		nodeQueryService:  nodeQueryService,
		graphQueryService: graphQueryService,
		eventBridgeClient: eventBridgeClient,
		container:         container,
	}
}

// CreateNode handles POST /api/nodes
// @Summary Create a new memory node
// @Description Creates a new memory node with content, optional title and tags. The system automatically extracts keywords and establishes connections to existing nodes.
// @Tags Memory Management
// @Accept json
// @Produce json
// @Security Bearer
// @Param request body api.CreateNodeRequest true "Memory node creation request"
// @Success 201 {object} api.Node "Successfully created memory node"
// @Failure 400 {object} api.ErrorResponse "Invalid request body or validation failed"
// @Failure 401 {object} api.ErrorResponse "Authentication required"
// @Failure 503 {object} api.ErrorResponse "Service temporarily unavailable"
// @Router /nodes [post]
func (h *MemoryHandler) CreateNode(w http.ResponseWriter, r *http.Request) {
	// Process create node request
	// Check if CQRS services are available
	if h.nodeService == nil {
		api.Error(w, http.StatusServiceUnavailable, "Service temporarily unavailable - CQRS migration in progress")
		return
	}

	userID, ok := getUserID(r)
	if !ok {
		api.Error(w, http.StatusUnauthorized, "Authentication required")
		return
	}
	var req api.CreateNodeRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		api.Error(w, http.StatusBadRequest, "Invalid request body")
		return
	}
	// Request decoded successfully

	if req.Content == "" {
		api.Error(w, http.StatusBadRequest, "Content cannot be empty")
		return
	}

	tags := req.Tags
	if tags == nil {
		tags = []string{}
	}

	// Create command for CQRS pattern
	cmd := &commands.CreateNodeCommand{
		UserID:  userID,
		Content: req.Content,
		Title:   req.Title,
		Tags:    tags,
	}
	// Command created for node creation

	// Add idempotency key if provided
	if idempotencyKey := r.Header.Get("Idempotency-Key"); idempotencyKey != "" {
		cmd.IdempotencyKey = idempotencyKey
	} else {
		// Generate automatic key
		cmd.IdempotencyKey = generateIdempotencyKey(userID, "CREATE_NODE", req)
	}

	// Execute command through CQRS service
	result, err := h.nodeService.CreateNode(r.Context(), cmd)
	if err != nil {
		handleServiceError(w, err)
		return
	}

	// Publish "NodeCreated" event to EventBridge with complete graph update
	eventDetail, err := json.Marshal(map[string]any{
		"type":      "nodeCreated",
		"userId":    result.Node.UserID,
		"nodeId":    result.Node.ID,
		"content":   result.Node.Content,
		"title":     result.Node.Title,
		"keywords":  result.Node.Keywords,
		"edges":     result.SuggestedEdges, // Use suggested edges from result
		"timestamp": time.Now(),
	})
	if err != nil {
		handleServiceError(w, err)
		return
	}

	_, err = h.eventBridgeClient.PutEvents(r.Context(), &eventbridge.PutEventsInput{
		Entries: []types.PutEventsRequestEntry{
			{
				Source:       aws.String("brain2.api"),
				DetailType:   aws.String("GraphUpdate"),
				Detail:       aws.String(string(eventDetail)),
				EventBusName: aws.String("B2EventBus"),
			},
		},
	})
	if err != nil {
		handleServiceError(w, err)
		return
	}

	api.Success(w, http.StatusCreated, api.NodeResponse{
		NodeID:    result.Node.ID,
		Content:   result.Node.Content,
		Title:     result.Node.Title,
		Tags:      result.Node.Tags,
		Timestamp: result.Node.CreatedAt.Format(time.RFC3339),
		Version:   result.Node.Version,
	})
}

// ListNodes handles GET /api/nodes
// @Summary List all memory nodes for the authenticated user
// @Description Retrieves all memory nodes belonging to the authenticated user, including content, tags, and metadata
// @Tags Memory Management
// @Produce json
// @Security Bearer
// @Success 200 {array} api.Node "Successfully retrieved user's memory nodes"
// @Failure 401 {object} api.ErrorResponse "Authentication required"
// @Failure 503 {object} api.ErrorResponse "Service temporarily unavailable"
// @Router /nodes [get]
func (h *MemoryHandler) ListNodes(w http.ResponseWriter, r *http.Request) {
	// Check if CQRS services are available
	if h.nodeQueryService == nil {
		api.Error(w, http.StatusServiceUnavailable, "Service temporarily unavailable - CQRS migration in progress")
		return
	}

	userID, ok := getUserID(r)
	if !ok {
		// Authentication failed
		api.Error(w, http.StatusUnauthorized, "Authentication required")
		return
	}

	// Parse query parameters
	urlQuery := r.URL.Query()
	limit := 20
	if l := urlQuery.Get("limit"); l != "" {
		if parsed, err := strconv.Atoi(l); err == nil && parsed > 0 && parsed <= 100 {
			limit = parsed
		}
	}

	// Processing list nodes request

	// Create query for CQRS pattern
	listQuery, err := queries.NewListNodesQuery(userID)
	if err != nil {
		// Failed to create query
		handleServiceError(w, err)
		return
	}

	listQuery.WithPagination(limit, urlQuery.Get("nextToken"))

	// Add search filter if provided
	if searchQuery := urlQuery.Get("search"); searchQuery != "" {
		listQuery.WithSearch(searchQuery)
	}

	// Add sorting if provided
	if sortBy := urlQuery.Get("sortBy"); sortBy != "" {
		sortDirection := urlQuery.Get("sortDirection")
		if sortDirection == "" {
			sortDirection = "desc"
		}
		listQuery.WithSort(sortBy, sortDirection)
	}

	response, err := h.nodeQueryService.ListNodes(r.Context(), listQuery)
	if err != nil {
		// Query service failed
		handleServiceError(w, err)
		return
	}

	if response == nil {
		// Received nil response from service
		api.Error(w, http.StatusInternalServerError, "Service returned no data")
		return
	}

	// Query executed successfully

	// Convert NodeView DTOs to API response format
	apiNodes := make([]api.Node, len(response.Nodes))
	for i, nodeView := range response.Nodes {
		apiNodes[i] = api.Node{
			NodeID:    nodeView.ID,
			UserID:    nodeView.UserID,
			Content:   nodeView.Content,
			Title:     nodeView.Title,
			Tags:      nodeView.Tags,
			Metadata:  nil, // NodeView doesn't include metadata field
			Timestamp: nodeView.CreatedAt.Format(time.RFC3339),
			CreatedAt: nodeView.CreatedAt.Format(time.RFC3339),
			UpdatedAt: nodeView.UpdatedAt.Format(time.RFC3339),
		}
	}

	// Response prepared successfully

	nodesResponse := map[string]interface{}{
		"nodes":     apiNodes,
		"total":     response.Total,
		"hasMore":   response.HasMore,
		"nextToken": response.NextToken,
	}

	// Response prepared successfully
	api.Success(w, http.StatusOK, nodesResponse)
}

// GetNode handles GET /api/nodes/{nodeId}
// @Summary Get a specific memory node by ID
// @Description Retrieves a specific memory node by its ID, including full details and connected edges
// @Tags Memory Management
// @Produce json
// @Security Bearer
// @Param nodeId path string true "Memory node ID"
// @Success 200 {object} api.NodeDetailsResponse "Successfully retrieved memory node details"
// @Failure 400 {object} api.ErrorResponse "Invalid node ID format"
// @Failure 401 {object} api.ErrorResponse "Authentication required"
// @Failure 404 {object} api.ErrorResponse "Memory node not found"
// @Failure 503 {object} api.ErrorResponse "Service temporarily unavailable"
// @Router /nodes/{nodeId} [get]
func (h *MemoryHandler) GetNode(w http.ResponseWriter, r *http.Request) {
	// Check if CQRS services are available
	if h.nodeQueryService == nil {
		api.Error(w, http.StatusServiceUnavailable, "Service temporarily unavailable - CQRS migration in progress")
		return
	}

	userID, ok := getUserID(r)
	if !ok {
		api.Error(w, http.StatusUnauthorized, "Authentication required")
		return
	}
	nodeID := chi.URLParam(r, "nodeId")

	// Check if this is a post-cold-start request for conditional behavior
	isPostColdStart := h.container.IsPostColdStartRequest()
	if isPostColdStart {
		timeSince := h.container.GetTimeSinceColdStart()
		// Processing post-cold-start request

		// Add cold start headers to response
		w.Header().Set("X-Cold-Start", "true")
		w.Header().Set("X-Cold-Start-Age", timeSince.String())
	}

	// Create query for CQRS pattern
	nodeQuery, err := queries.NewGetNodeQuery(userID, nodeID)
	if err != nil {
		if isPostColdStart {
			// Error creating query during post-cold-start request
		}
		handleServiceError(w, err)
		return
	}

	// Include connections in the query
	nodeQuery.WithConnections()

	result, err := h.nodeQueryService.GetNode(r.Context(), nodeQuery)
	if err != nil {
		if isPostColdStart {
			// Error during post-cold-start request
		}
		handleServiceError(w, err)
		return
	}

	// Extract connected node IDs from connections
	connectedNodeIDs := make(map[string]bool) // Use map to avoid duplicates
	for _, connection := range result.Connections {
		// Add the "other" node ID
		if connection.SourceNodeID == nodeID {
			// Current node is source, so target is the connected node
			if connection.TargetNodeID != nodeID { // Avoid self-references
				connectedNodeIDs[connection.TargetNodeID] = true
			}
		} else if connection.TargetNodeID == nodeID {
			// Current node is target, so source is the connected node
			if connection.SourceNodeID != nodeID { // Avoid self-references
				connectedNodeIDs[connection.SourceNodeID] = true
			}
		}
	}

	// Convert map to slice
	connectedNodes := make([]string, 0, len(connectedNodeIDs))
	for id := range connectedNodeIDs {
		connectedNodes = append(connectedNodes, id)
	}

	response := api.NodeDetailsResponse{
		NodeID:    result.Node.ID,
		Content:   result.Node.Content,
		Title:     result.Node.Title,
		Tags:      result.Node.Tags,
		Timestamp: result.Node.CreatedAt.Format(time.RFC3339),
		Version:   result.Node.Version,
		Edges:     connectedNodes, // Edges field contains connected node IDs, not edge IDs
	}

	if isPostColdStart {
		log.Printf("GetNode: Successfully completed post-cold-start request for node %s", nodeID)
	}

	api.Success(w, http.StatusOK, response)
}

// UpdateNode handles PUT /api/nodes/{nodeId}
// @Summary Update an existing memory node
// @Description Updates an existing memory node's content, title, and tags. The system will re-analyze connections.
// @Tags Memory Management
// @Accept json
// @Produce json
// @Security Bearer
// @Param nodeId path string true "Memory node ID"
// @Param request body api.UpdateNodeRequest true "Memory node update request"
// @Success 200 {object} api.NodeResponse "Successfully updated memory node"
// @Failure 400 {object} api.ErrorResponse "Invalid request body or node ID"
// @Failure 401 {object} api.ErrorResponse "Authentication required"
// @Failure 404 {object} api.ErrorResponse "Memory node not found"
// @Failure 503 {object} api.ErrorResponse "Service temporarily unavailable"
// @Router /nodes/{nodeId} [put]
func (h *MemoryHandler) UpdateNode(w http.ResponseWriter, r *http.Request) {
	// Check if CQRS services are available
	if h.nodeService == nil {
		api.Error(w, http.StatusServiceUnavailable, "Service temporarily unavailable - CQRS migration in progress")
		return
	}

	userID, ok := getUserID(r)
	if !ok {
		api.Error(w, http.StatusUnauthorized, "Authentication required")
		return
	}
	nodeID := chi.URLParam(r, "nodeId")

	var req api.UpdateNodeRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		api.Error(w, http.StatusBadRequest, "Invalid request body")
		return
	}

	// Add server-side validation
	if req.Content != "" && len(req.Content) > 5000 {
		api.Error(w, http.StatusBadRequest, "Content must not exceed 5000 characters.")
		return
	}

	tags := req.Tags
	if tags == nil {
		tags = []string{}
	}

	// Create command for CQRS pattern
	cmd := &commands.UpdateNodeCommand{
		NodeID:  nodeID,
		UserID:  userID,
		Content: req.Content,
		Title:   req.Title,
		Tags:    tags,
	}

	// Add idempotency key if provided
	if idempotencyKey := r.Header.Get("Idempotency-Key"); idempotencyKey != "" {
		// Note: UpdateNodeCommand doesn't have IdempotencyKey field,
		// but we'll handle it in the service layer if needed
	}

	_, err := h.nodeService.UpdateNode(r.Context(), cmd)
	if err != nil {
		handleServiceError(w, err)
		return
	}

	api.Success(w, http.StatusOK, map[string]string{"message": "Node updated successfully"})
}

// DeleteNode handles DELETE /api/nodes/{nodeId}
// @Summary Delete a memory node
// @Description Deletes a memory node and all its connections. This operation cannot be undone.
// @Tags Memory Management
// @Security Bearer
// @Param nodeId path string true "Memory node ID"
// @Success 204 "Memory node successfully deleted"
// @Failure 401 {object} api.ErrorResponse "Authentication required"
// @Failure 404 {object} api.ErrorResponse "Memory node not found"
// @Failure 503 {object} api.ErrorResponse "Service temporarily unavailable"
// @Router /nodes/{nodeId} [delete]
func (h *MemoryHandler) DeleteNode(w http.ResponseWriter, r *http.Request) {
	// Check if CQRS services are available
	if h.nodeService == nil {
		api.Error(w, http.StatusServiceUnavailable, "Service temporarily unavailable - CQRS migration in progress")
		return
	}

	userID, ok := getUserID(r)
	if !ok {
		api.Error(w, http.StatusUnauthorized, "Authentication required")
		return
	}
	nodeID := chi.URLParam(r, "nodeId")

	// Create command for CQRS pattern
	cmd := &commands.DeleteNodeCommand{
		NodeID: nodeID,
		UserID: userID,
	}

	_, err := h.nodeService.DeleteNode(r.Context(), cmd)
	if err != nil {
		handleServiceError(w, err)
		return
	}

	api.Success(w, http.StatusOK, map[string]string{"message": "Node deleted successfully"})
}

// BulkDeleteNodes handles POST /api/nodes/bulk-delete
// @Summary Delete multiple memory nodes
// @Description Deletes multiple memory nodes and all their connections in a single operation
// @Tags Memory Management
// @Accept json
// @Produce json
// @Security Bearer
// @Param request body api.BulkDeleteRequest true "Bulk delete request with node IDs"
// @Success 200 {object} api.BulkDeleteResponse "Bulk delete operation completed"
// @Failure 400 {object} api.ErrorResponse "Invalid request body"
// @Failure 401 {object} api.ErrorResponse "Authentication required"
// @Failure 503 {object} api.ErrorResponse "Service temporarily unavailable"
// @Router /nodes/bulk-delete [post]
func (h *MemoryHandler) BulkDeleteNodes(w http.ResponseWriter, r *http.Request) {
	// Check if CQRS services are available
	if h.nodeService == nil {
		api.Error(w, http.StatusServiceUnavailable, "Service temporarily unavailable - CQRS migration in progress")
		return
	}

	userID, ok := getUserID(r)
	if !ok {
		api.Error(w, http.StatusUnauthorized, "Authentication required")
		return
	}

	var req api.BulkDeleteRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		api.Error(w, http.StatusBadRequest, "Invalid request body")
		return
	}

	if len(req.NodeIDs) == 0 {
		api.Error(w, http.StatusBadRequest, "NodeIds cannot be empty")
		return
	}

	if len(req.NodeIDs) > 100 {
		api.Error(w, http.StatusBadRequest, "Cannot delete more than 100 nodes at once")
		return
	}

	// Create command for CQRS pattern
	cmd := &commands.BulkDeleteNodesCommand{
		NodeIDs: req.NodeIDs,
		UserID:  userID,
	}

	result, err := h.nodeService.BulkDeleteNodes(r.Context(), cmd)
	if err != nil {
		handleServiceError(w, err)
		return
	}

	message := fmt.Sprintf("Successfully deleted %d nodes", result.DeletedCount)
	if len(result.FailedIDs) > 0 {
		message += fmt.Sprintf(", failed to delete %d nodes", len(result.FailedIDs))
	}

	api.Success(w, http.StatusOK, api.BulkDeleteResponse{
		DeletedCount: result.DeletedCount,
		FailedIDs:    result.FailedIDs,
	})
}

// GetGraphData handles GET /api/graph-data
// @Summary Get graph visualization data
// @Description Retrieves graph data for visualization including nodes and edges in a format suitable for graph libraries
// @Tags Graph Operations
// @Produce json
// @Security Bearer
// @Success 200 {object} api.GraphDataResponse "Graph data for visualization"
// @Failure 401 {object} api.ErrorResponse "Authentication required"
// @Failure 503 {object} api.ErrorResponse "Service temporarily unavailable"
// @Router /graph-data [get]
func (h *MemoryHandler) GetGraphData(w http.ResponseWriter, r *http.Request) {
	// Check if CQRS services are available
	if h.graphQueryService == nil {
		api.Error(w, http.StatusServiceUnavailable, "Service temporarily unavailable - CQRS migration in progress")
		return
	}

	userID, ok := getUserID(r)
	if !ok {
		log.Printf("ERROR: GetGraphData - Authentication required, getUserID returned false")
		api.Error(w, http.StatusUnauthorized, "Authentication required")
		return
	}

	log.Printf("DEBUG: GetGraphData called for userID: %s", userID)

	// Create query for CQRS pattern
	graphQuery := &queries.GetGraphQuery{
		UserID:         userID,
		Limit:          0, // No limit for now
		IncludeMetrics: false,
	}

	log.Printf("DEBUG: Calling graphQueryService.GetGraph")
	result, err := h.graphQueryService.GetGraph(r.Context(), graphQuery)
	if err != nil {
		log.Printf("ERROR: GetGraphData failed: %v", err)
		handleServiceError(w, err)
		return
	}

	var elements []api.GraphDataResponse_Elements_Item

	// Handle case where result is nil (should not happen, but defensive programming)
	if result == nil {
		log.Printf("WARN: GetGraphData returned nil result, returning empty response")
		api.Success(w, http.StatusOK, api.GraphDataResponse{Elements: &elements})
		return
	}

	log.Printf("DEBUG: GetGraphData succeeded, graph has %d nodes and %d edges", len(result.Nodes), len(result.Edges))

	for _, nodeView := range result.Nodes {
		label := nodeView.Title
		if label == "" {
			label = nodeView.Content
		}
		if len(label) > 50 {
			label = label[:47] + "..."
		}

		graphNode := api.GraphNode{
			Data: &api.NodeData{
				Id:    nodeView.ID,
				Label: label,
			},
		}

		var element api.GraphDataResponse_Elements_Item
		if err := element.FromGraphNode(graphNode); err != nil {
			log.Printf("WARN: Failed to convert node to GraphDataResponse element: %v", err)
			continue
		}
		elements = append(elements, element)
	}

	for _, edgeView := range result.Edges {
		edgeID := fmt.Sprintf("%s-%s", edgeView.SourceID, edgeView.TargetID)
		graphEdge := api.GraphEdge{
			Data: &api.EdgeData{
				Id:     edgeID,
				Source: edgeView.SourceID,
				Target: edgeView.TargetID,
			},
		}

		var element api.GraphDataResponse_Elements_Item
		if err := element.FromGraphEdge(graphEdge); err != nil {
			log.Printf("WARN: Failed to convert edge to GraphDataResponse element: %v", err)
			continue
		}
		elements = append(elements, element)
	}

	log.Printf("DEBUG: GetGraphData completed successfully - returning %d elements", len(elements))
	api.Success(w, http.StatusOK, api.GraphDataResponse{Elements: &elements})
}

// checkOwnership is no longer needed as CQRS services handle ownership internally

// generateIdempotencyKey creates an idempotency key from request data if not provided
func generateIdempotencyKey(userID, operation string, payload interface{}) string {
	hasher := sha256.New()

	// Include user ID and operation in the hash
	hasher.Write([]byte(userID))
	hasher.Write([]byte(operation))

	// Include payload in the hash
	if payload != nil {
		payloadBytes, err := json.Marshal(payload)
		if err == nil {
			hasher.Write(payloadBytes)
		}
	}

	return hex.EncodeToString(hasher.Sum(nil))
}

// getIdempotencyKey extracts idempotency key from header or generates one
func getIdempotencyKey(r *http.Request, userID, operation string, payload interface{}) string {
	// Check for client-provided idempotency key
	if key := r.Header.Get("Idempotency-Key"); key != "" {
		return key
	}

	// Generate automatic key based on operation and payload
	return generateIdempotencyKey(userID, operation, payload)
}