# backend/pkg/api/swagger.yaml
basePath: /api
definitions:
  brain2-backend_pkg_api.BulkDeleteRequest:
    properties:
      nodeIds:
        items:
          type: string
        type: array
    type: object
  brain2-backend_pkg_api.BulkDeleteResponse:
    properties:
      deletedCount:
        type: integer
      failedIds:
        items:
          type: string
        type: array
    type: object
...
  brain2-backend_pkg_api.CreateNodeRequest:
    properties:
      content:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
    type: object
  brain2-backend_pkg_api.ErrorResponse:
    properties:
      error:
        example: Invalid request body
        type: string
    type: object
  brain2-backend_pkg_api.GraphDataResponse:
    properties:
      elements:
        items:
          $ref: '#/definitions/brain2-backend_pkg_api.GraphDataResponse_Elements_Item'
        type: array
    type: object
  brain2-backend_pkg_api.GraphDataResponse_Elements_Item:
    properties:
      data: {}
      group:
        type: string
    type: object
...
  brain2-backend_pkg_api.Node:
    properties:
      content:
        type: string
      createdAt:
        type: string
      metadata:
        additionalProperties: true
        type: object
      nodeId:
        type: string
      tags:
        items:
          type: string
        type: array
      timestamp:
        type: string
      title:
        type: string
      updatedAt:
        type: string
      user_id:
        type: string
    type: object
  brain2-backend_pkg_api.NodeDetailsResponse:
    properties:
      content:
        type: string
      edges:
        items:
          type: string
        type: array
      nodeId:
        type: string
      tags:
        items:
          type: string
        type: array
      timestamp:
        type: string
      title:
        type: string
      version:
        type: integer
    type: object
  brain2-backend_pkg_api.NodeResponse:
    properties:
      content:
        type: string
      nodeId:
        type: string
      tags:
        items:
          type: string
        type: array
      timestamp:
        type: string
      title:
        type: string
      version:
        type: integer
    type: object
...
  brain2-backend_pkg_api.UpdateNodeRequest:
    properties:
      content:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
    type: object
host: api.brain2.example.com
info:
  contact:
    email: support@brain2.example.com
    name: Brain2 API Support
    url: https://github.com/your-org/brain2-backend
  description: |-
    A RESTful API for managing personal knowledge graphs with automatic memory connections and intelligent categorization.

    ## Overview
    The Brain2 API allows users to create, manage, and explore their personal knowledge graphs.
    It provides endpoints for managing memory nodes, categories, and graph relationships with
    automatic connection discovery and intelligent categorization features.

    ## Authentication
    All API endpoints require JWT authentication via the Authorization header.

    ## Rate Limiting
    API requests are rate-limited per user to ensure fair usage and system stability.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Brain2 Knowledge Graph API
  version: 1.0.0
paths:
...
  /graph-data:
    get:
      description: Retrieves graph data for visualization including nodes and edges
        in a format suitable for graph libraries
      produces:
      - application/json
      responses:
        "200":
          description: Graph data for visualization
          schema:
            $ref: '#/definitions/brain2-backend_pkg_api.GraphDataResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/brain2-backend_pkg_api.ErrorResponse'
        "503":
          description: Service temporarily unavailable
          schema:
            $ref: '#/definitions/brain2-backend_pkg_api.ErrorResponse'
      security:
      - Bearer: []
      summary: Get graph visualization data
      tags:
      - Graph Operations
...
  /nodes:
    get:
      description: Retrieves all memory nodes belonging to the authenticated user,
        including content, tags, and metadata
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved user's memory nodes
          schema:
            items:
              $ref: '#/definitions/brain2-backend_pkg_api.Node'
            type: array
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/brain2-backend_pkg_api.ErrorResponse'
        "503":
          description: Service temporarily unavailable
          schema:
            $ref: '#/definitions/brain2-backend_pkg_api.ErrorResponse'
      security:
      - Bearer: []
      summary: List all memory nodes for the authenticated user
      tags:
      - Memory Management
    post:
      consumes:
      - application/json
      description: Creates a new memory node with content, optional title and tags.
        The system automatically extracts keywords and establishes connections to
        existing nodes.
      parameters:
      - description: Memory node creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/brain2-backend_pkg_api.CreateNodeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created memory node
          schema:
            $ref: '#/definitions/brain2-backend_pkg_api.Node'
        "400":
          description: Invalid request body or validation failed
          schema:
            $ref: '#/definitions/brain2-backend_pkg_api.ErrorResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/brain2-backend_pkg_api.ErrorResponse'
        "503":
          description: Service temporarily unavailable
          schema:
            $ref: '#/definitions/brain2-backend_pkg_api.ErrorResponse'
      security:
      - Bearer: []
      summary: Create a new memory node
      tags:
      - Memory Management
  /nodes/{nodeId}:
    delete:
      description: Deletes a memory node and all its connections. This operation cannot
        be undone.
      parameters:
      - description: Memory node ID
        in: path
        name: nodeId
        required: true
        type: string
      responses:
        "204":
          description: Memory node successfully deleted
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/brain2-backend_pkg_api.ErrorResponse'
        "404":
          description: Memory node not found
          schema:
            $ref: '#/definitions/brain2-backend_pkg_api.ErrorResponse'
        "503":
          description: Service temporarily unavailable
          schema:
            $ref: '#/definitions/brain2-backend_pkg_api.ErrorResponse'
      security:
      - Bearer: []
      summary: Delete a memory node
      tags:
      - Memory Management
    get:
      description: Retrieves a specific memory node by its ID, including full details
        and connected edges
      parameters:
      - description: Memory node ID
        in: path
        name: nodeId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved memory node details
          schema:
            $ref: '#/definitions/brain2-backend_pkg_api.NodeDetailsResponse'
        "400":
          description: Invalid node ID format
          schema:
            $ref: '#/definitions/brain2-backend_pkg_api.ErrorResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/brain2-backend_pkg_api.ErrorResponse'
        "404":
          description: Memory node not found
          schema:
            $ref: '#/definitions/brain2-backend_pkg_api.ErrorResponse'
        "503":
          description: Service temporarily unavailable
          schema:
            $ref: '#/definitions/brain2-backend_pkg_api.ErrorResponse'
      security:
      - Bearer: []
      summary: Get a specific memory node by ID
      tags:
      - Memory Management
    put:
      consumes:
      - application/json
      description: Updates an existing memory node's content, title, and tags. The
        system will re-analyze connections.
      parameters:
      - description: Memory node ID
        in: path
        name: nodeId
        required: true
        type: string
      - description: Memory node update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/brain2-backend_pkg_api.UpdateNodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated memory node
          schema:
            $ref: '#/definitions/brain2-backend_pkg_api.NodeResponse'
        "400":
          description: Invalid request body or node ID
          schema:
            $ref: '#/definitions/brain2-backend_pkg_api.ErrorResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/brain2-backend_pkg_api.ErrorResponse'
        "404":
          description: Memory node not found
          schema:
            $ref: '#/definitions/brain2-backend_pkg_api.ErrorResponse'
        "503":
          description: Service temporarily unavailable
          schema:
            $ref: '#/definitions/brain2-backend_pkg_api.ErrorResponse'
      security:
      - Bearer: []
      summary: Update an existing memory node
      tags:
      - Memory Management
  /nodes/bulk-delete:
    post:
      consumes:
      - application/json
      description: Deletes multiple memory nodes and all their connections in a single
        operation
      parameters:
      - description: Bulk delete request with node IDs
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/brain2-backend_pkg_api.BulkDeleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Bulk delete operation completed
          schema:
            $ref: '#/definitions/brain2-backend_pkg_api.BulkDeleteResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/brain2-backend_pkg_api.ErrorResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/brain2-backend_pkg_api.ErrorResponse'
        "503":
          description: Service temporarily unavailable
          schema:
            $ref: '#/definitions/brain2-backend_pkg_api.ErrorResponse'
      security:
      - Bearer: []
      summary: Delete multiple memory nodes
      tags:
      - Memory Management
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"